+++ tests/integ/test_data_source_api.py
@@ -64,7 +64,6 @@ from tests.resources.test_data_source_dir.test_data_source_data import (
     SQLITE3_DB_CUSTOM_SCHEMA_STRING,
     SQLITE3_DB_CUSTOM_SCHEMA_STRUCT_TYPE,
     sql_server_udtf_ingestion_data,
-    create_connection_oracledb,
     sql_server_create_connection_unicode_data,
     sql_server_create_connection_double_quoted_data,
 )
@@ -916,23 +915,6 @@ def test_fetch_merge_count_integ(session, caplog):
             assert df.order_by("ID").collect() == assert_data
             # 2 batch + 2 fetch size = 2 parquet file
             assert caplog.text.count("Retrieved file from parquet queue") == 2
-
-
-def test_unicode_column_name_oracledb(session):
-    df = session.read.dbapi(create_connection_oracledb, table='"用户資料"')
-    assert df.collect() == [Row(編號=1, 姓名="山田太郎", 國家="日本", 備註="これはUnicodeテストです")]
-
-
-def test_double_quoted_column_name_oracledb(session):
-    df = session.read.dbapi(create_connection_oracledb, table='"UserProfile"')
-    assert df.collect() == [
-        Row(
-            Id=1,
-            FullName="John Doe",
-            Country="USA",
-            Notes="This is a case-sensitive example.",
-        )
-    ]
 
 
 def test_unicode_column_name_sql_server(session):
